Move TSData class so it can be seen from ts_lowering_context without changes to BackendData inferface
